+++ Changes.md
@@ -1,3 +1,4 @@
+* #1222 - Display all performance runs in the performance runs vs. schedule report
 * Fix bug in subjective application where the score summary sometimes would not display
 * #1211 - Add PDF version of the schedule
 * #1224 - allow judge names to be longer than 64 characters